fix: some bugs
