+++ swanlab/data/modules/__init__.py
@@ -3,7 +3,7 @@ from typing import List, Union
 from swankit.core.data import BaseType, MediaBuffer, MediaType
 
 from .audio import Audio
-from .echarts import echarts, Echarts, PyEchartsBase,PyEchartsTable
+from .echarts import echarts, Echarts, PyEchartsBase, PyEchartsTable
 from .image import Image
 from .line import FloatConvertible, Line
 from .object3d import Model3D, Object3D, PointCloud, Molecule
@@ -31,5 +31,5 @@ __all__ = [
     "Echarts",
     "echarts",
     "PyEchartsBase",
-    "PyEchartsTable"
+    "PyEchartsTable",
 ]
