+++ swanlab/data/modules/echarts/echarts.py
@@ -1,0 +1,43 @@
+"""
+@author: cunyue
+@file: echarts.py
+@time: 2025/6/11 11:37
+@description: 由于前端支持有限，这里仅导出部分 echarts 类型
+"""
+
+from pyecharts import options as opts
+from pyecharts.charts import *
+
+from .table import Table
+
+__all__ = [
+    "opts",
+    "Bar3D",
+    "Bar",
+    "Boxplot",
+    "Calendar",
+    "Kline",
+    "Grid",
+    "Scatter",
+    "EffectScatter",
+    "Funnel",
+    "Gauge",
+    "Graph",
+    "HeatMap",
+    "Line",
+    "Line3D",
+    "Liquid",
+    "Parallel",
+    "PictorialBar",
+    "Pie",
+    "Polar",
+    "Radar",
+    "Sankey",
+    "Scatter3D",
+    "Sunburst",
+    "Surface3D",
+    "ThemeRiver",
+    "Tree",
+    "TreeMap",
+    "Table",
+]
