+++ swanlab/data/sdk.py
@@ -18,7 +18,7 @@ from swanlab.log import swanlog
 from swanlab.swanlab_settings import Settings, get_settings, set_settings
 from .callbacker.cloud import CloudRunCallback
 from .formatter import check_load_json_yaml, check_callback_format
-from .modules import DataType, PyEchartsBase
+from .modules import DataType, PyEchartsTable, PyEchartsBase
 from .run import (
     SwanLabRunState,
     SwanLabRun,
@@ -249,7 +249,7 @@ register_callbacks = initializer.register_callbacks
 
 @should_call_after_init("You must call swanlab.init() before using log()")
 def log(
-    data: Dict[str, Union[DataType, PyEchartsBase, List[PyEchartsBase]]],
+    data: Dict[str, Union[DataType, PyEchartsBase, PyEchartsTable, List[PyEchartsTable], List[PyEchartsBase]]],
     step: int = None,
     print_to_console: bool = False,
 ):
