+++ test/metrics/echarts/README.md
@@ -1,9 +1,12 @@
-# echarts测试需知
+# echarts测试
 
-# NOTE 你可能需要下载对应的文件才可以运行测试
+这里是根据 pyecharts-gallery 中的 demo, 对 pyecharts 模块的遍历测试，不保证 swanlab 能够完全渲染所有的图表。
 
----
+> 如果你有相关图表的需求，可以向我们提交 [issue](https://github.com/SwanHubX/SwanLab/issues)
+
 ## 需要下载额外文件的测试
+
+> NOTE 你可能需要下载对应的文件才可以运行测试
 
 - `dataset.py`测试:需要到[这里](https://github.com/pyecharts/pyecharts-gallery/blob/master/Dataset/life-expectancy-table.json),下载文件`life-expectancy-table.json`到本目录下的`assets/echarts/dataset`目录下,才能运行`dataset.py`测试
 - `graph.py`测试:需要到[这里](https://github.com/pyecharts/pyecharts-gallery/blob/master/Graph/les-miserables.json),[这里](https://github.com/pyecharts/pyecharts-gallery/blob/master/Graph/weibo.json),下载两个文件`les-miserables.json`,`weibo.json`到本目录下的`assets/echarts/graph`目录下,才能运行`graph.py`测试
